+++ feature/preview/src/main/java/com/google/jetpackcamera/feature/preview/PreviewViewModel.kt
@@ -43,20 +43,20 @@ import com.google.jetpackcamera.settings.model.applyExternalCaptureMode
 import com.google.jetpackcamera.ui.components.capture.LOW_LIGHT_BOOST_FAILURE_TAG
 import com.google.jetpackcamera.ui.components.capture.R
 import com.google.jetpackcamera.ui.components.capture.ScreenFlash
-import com.google.jetpackcamera.ui.components.capture.SnackBarController
-import com.google.jetpackcamera.ui.components.capture.SnackBarControllerImpl
-import com.google.jetpackcamera.ui.components.capture.controller.CameraController
-import com.google.jetpackcamera.ui.components.capture.controller.CameraControllerImpl
-import com.google.jetpackcamera.ui.components.capture.controller.CaptureController
-import com.google.jetpackcamera.ui.components.capture.controller.CaptureControllerImpl
-import com.google.jetpackcamera.ui.components.capture.controller.ImageWellController
-import com.google.jetpackcamera.ui.components.capture.controller.ImageWellControllerImpl
-import com.google.jetpackcamera.ui.components.capture.controller.ZoomController
-import com.google.jetpackcamera.ui.components.capture.controller.ZoomControllerImpl
-import com.google.jetpackcamera.ui.components.capture.debug.controller.DebugController
-import com.google.jetpackcamera.ui.components.capture.debug.controller.DebugControllerImpl
-import com.google.jetpackcamera.ui.components.capture.quicksettings.controller.QuickSettingsController
-import com.google.jetpackcamera.ui.components.capture.quicksettings.controller.QuickSettingsControllerImpl
+import com.google.jetpackcamera.ui.controller.CameraController
+import com.google.jetpackcamera.ui.controller.CaptureController
+import com.google.jetpackcamera.ui.controller.ImageWellController
+import com.google.jetpackcamera.ui.controller.SnackBarController
+import com.google.jetpackcamera.ui.controller.ZoomController
+import com.google.jetpackcamera.ui.controller.debug.DebugController
+import com.google.jetpackcamera.ui.controller.impl.CameraControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.CaptureControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.DebugControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.ImageWellControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.QuickSettingsControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.SnackBarControllerImpl
+import com.google.jetpackcamera.ui.controller.impl.ZoomControllerImpl
+import com.google.jetpackcamera.ui.controller.quicksettings.QuickSettingsController
 import com.google.jetpackcamera.ui.uistate.SnackBarUiState
 import com.google.jetpackcamera.ui.uistate.SnackbarData
 import com.google.jetpackcamera.ui.uistate.capture.DebugUiState
@@ -66,7 +66,6 @@ import com.google.jetpackcamera.ui.uistateadapter.capture.compound.captureUiStat
 import com.google.jetpackcamera.ui.uistateadapter.capture.debugUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
-import kotlin.coroutines.coroutineContext
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
