+++ settings.gradle.kts
@@ -56,3 +56,5 @@ include(":core:model")
 include(":ui:uistate:postcapture")
 include(":ui:uistateadapter:postcapture")
 include(":core:camera:postprocess")
+include(":ui:controller")
+include(":ui:controller:impl")
