+++ ui/controller/impl/src/main/java/com/google/jetpackcamera/ui/controller/impl/ZoomControllerImpl.kt
@@ -13,40 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.ui.components.capture.debug.controller
+package com.google.jetpackcamera.ui.controller.impl
 
 import com.google.jetpackcamera.core.camera.CameraSystem
-import com.google.jetpackcamera.model.TestPattern
+import com.google.jetpackcamera.model.CameraZoomRatio
+import com.google.jetpackcamera.ui.controller.ZoomController
 import com.google.jetpackcamera.ui.uistate.capture.TrackedCaptureUiState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 
 /**
- * Implementation of [DebugController] that interacts with [CameraSystem] and updates
- * [trackedCaptureUiState].
+ * Implementation of [ZoomController] that updates the camera's zoom and tracked UI state.
  *
- * @param cameraSystem The camera system to control.
- * @param trackedCaptureUiState The state flow to update with debug information.
+ * @param cameraSystem The camera system to update zoom on.
+ * @param trackedCaptureUiState State for tracking zoom changes.
  */
-class DebugControllerImpl(
+class ZoomControllerImpl(
     private val cameraSystem: CameraSystem,
     private val trackedCaptureUiState: MutableStateFlow<TrackedCaptureUiState>
-) : DebugController {
-    override fun toggleDebugHidingComponents() {
-        trackedCaptureUiState.update { old ->
-            old.copy(debugHidingComponents = !old.debugHidingComponents)
-        }
-    }
+) : ZoomController {
 
-    override fun toggleDebugOverlay() {
-        trackedCaptureUiState.update { old ->
-            old.copy(isDebugOverlayOpen = !old.isDebugOverlayOpen)
-        }
-    }
-
-    override fun setTestPattern(testPattern: TestPattern) {
-        cameraSystem.setTestPattern(
-            newTestPattern = testPattern
+    override fun setZoomRatio(zoomRatio: CameraZoomRatio) {
+        cameraSystem.changeZoomRatio(
+            newZoomState = zoomRatio
         )
+    }
+
+    override fun setZoomAnimationState(targetValue: Float?) {
+        trackedCaptureUiState.update { old ->
+            old.copy(zoomAnimationTarget = targetValue)
+        }
     }
 }
