+++ ui/controller/src/main/java/com/google/jetpackcamera/ui/controller/CameraController.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.ui.components.capture.controller
+package com.google.jetpackcamera.ui.controller
 
 import com.google.jetpackcamera.model.DeviceRotation
